requirements.txt
pytest==5.4.1
Pillow=8.3.1
